pyproject.toml
[project]
name = "iaaa25-rfe-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "catboost>=1.2.8",
    "gdown>=5.2.1",
    "imbalanced-learn>=0.14.1",
    "imblearn>=0.0",
    "keras-tuner>=1.4.8",
    "lightgbm>=4.6.0",
    "nolds>=0.6.3",
    "numba>=0.63.1",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "optuna>=4.6.0",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "pexpect>=4.9.0",
    "pycatch22>=0.4.5",
    "pyts>=0.13.0",
    "saxpy>=1.0.1.dev167",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seglearn>=1.2.5",
    "tensorflow>=2.20.0",
    "tsassure-feature>=0.7",
    "tsfresh>=0.21.1",
    "tslearn>=0.7.0",
    "xgboost>=3.1.3",
]
